Fix MC Dropout: use CUDA, float32, eval mode for backbone, T=20 sampling

+++ sam3/model/vitdet.py
@@ -71,7 +71,7 @@ class Mlp(nn.Module):
         x = addmm_act(type(self.act), self.fc1, x)
         x = self.drop1(x)
         x = self.norm(x)
-        x = self.fc2(x)
+        x = self.fc2(x.to(self.fc2.weight.dtype)) 
         x = self.drop2(x)
         return x
